+++ test.py
@@ -10,14 +10,14 @@ from urllib.parse import parse_qsl, urlparse
 from warnings import warn
 
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), "src"))
-import pyotp  # noqa
+import rlotp  # noqa
 
 
 class HOTPExampleValuesFromTheRFC(unittest.TestCase):
     def test_match_rfc(self):
         # 12345678901234567890 in Bas32
         # GEZDGNBVGY3TQOJQGEZDGNBVGY3TQOJQ
-        hotp = pyotp.HOTP("GEZDGNBVGY3TQOJQGEZDGNBVGY3TQOJQ")
+        hotp = rlotp.HOTP("GEZDGNBVGY3TQOJQGEZDGNBVGY3TQOJQ")
         self.assertEqual(hotp.at(0), "755224")
         self.assertEqual(hotp.at(1), "287082")
         self.assertEqual(hotp.at(2), "359152")
@@ -30,35 +30,35 @@ class HOTPExampleValuesFromTheRFC(unittest.TestCase):
         self.assertEqual(hotp.at(9), "520489")
 
     def test_invalid_input(self):
-        hotp = pyotp.HOTP("GEZDGNBVGY3TQOJQGEZDGNBVGY3TQOJQ")
+        hotp = rlotp.HOTP("GEZDGNBVGY3TQOJQGEZDGNBVGY3TQOJQ")
         with self.assertRaises(ValueError):
             hotp.at(-1)
 
     def test_verify_otp_reuse(self):
-        hotp = pyotp.HOTP("GEZDGNBVGY3TQOJQGEZDGNBVGY3TQOJQ")
+        hotp = rlotp.HOTP("GEZDGNBVGY3TQOJQGEZDGNBVGY3TQOJQ")
         self.assertTrue(hotp.verify("520489", 9))
         self.assertFalse(hotp.verify("520489", 10))
         self.assertFalse(hotp.verify("520489", 10))
 
     def test_provisioning_uri(self):
-        hotp = pyotp.HOTP("wrn3pqx5uqxqvnqr", name="mark@percival")
+        hotp = rlotp.HOTP("wrn3pqx5uqxqvnqr", name="mark@percival")
 
         url = urlparse(hotp.provisioning_uri())
         self.assertEqual(url.scheme, "otpauth")
         self.assertEqual(url.netloc, "hotp")
         self.assertEqual(url.path, "/mark%40percival")
         self.assertEqual(dict(parse_qsl(url.query)), {"secret": "wrn3pqx5uqxqvnqr", "counter": "0"})
-        self.assertEqual(hotp.provisioning_uri(), pyotp.parse_uri(hotp.provisioning_uri()).provisioning_uri())
+        self.assertEqual(hotp.provisioning_uri(), rlotp.parse_uri(hotp.provisioning_uri()).provisioning_uri())
 
-        hotp = pyotp.HOTP("wrn3pqx5uqxqvnqr", name="mark@percival", initial_count=12)
+        hotp = rlotp.HOTP("wrn3pqx5uqxqvnqr", name="mark@percival", initial_count=12)
         url = urlparse(hotp.provisioning_uri())
         self.assertEqual(url.scheme, "otpauth")
         self.assertEqual(url.netloc, "hotp")
         self.assertEqual(url.path, "/mark%40percival")
         self.assertEqual(dict(parse_qsl(url.query)), {"secret": "wrn3pqx5uqxqvnqr", "counter": "12"})
-        self.assertEqual(hotp.provisioning_uri(), pyotp.parse_uri(hotp.provisioning_uri()).provisioning_uri())
+        self.assertEqual(hotp.provisioning_uri(), rlotp.parse_uri(hotp.provisioning_uri()).provisioning_uri())
 
-        hotp = pyotp.HOTP("wrn3pqx5uqxqvnqr", name="mark@percival", issuer="FooCorp!")
+        hotp = rlotp.HOTP("wrn3pqx5uqxqvnqr", name="mark@percival", issuer="FooCorp!")
         url = urlparse(hotp.provisioning_uri())
         self.assertEqual(url.scheme, "otpauth")
         self.assertEqual(url.netloc, "hotp")
@@ -66,10 +66,10 @@ class HOTPExampleValuesFromTheRFC(unittest.TestCase):
         self.assertEqual(
             dict(parse_qsl(url.query)), {"secret": "wrn3pqx5uqxqvnqr", "counter": "0", "issuer": "FooCorp!"}
         )
-        self.assertEqual(hotp.provisioning_uri(), pyotp.parse_uri(hotp.provisioning_uri()).provisioning_uri())
+        self.assertEqual(hotp.provisioning_uri(), rlotp.parse_uri(hotp.provisioning_uri()).provisioning_uri())
 
         key = "c7uxuqhgflpw7oruedmglbrk7u6242vb"
-        hotp = pyotp.HOTP(key, digits=8, digest=hashlib.sha256, name="baco@peperina", issuer="FooCorp")
+        hotp = rlotp.HOTP(key, digits=8, digest=hashlib.sha256, name="baco@peperina", issuer="FooCorp")
         url = urlparse(hotp.provisioning_uri())
         self.assertEqual(url.scheme, "otpauth")
         self.assertEqual(url.netloc, "hotp")
@@ -84,9 +84,9 @@ class HOTPExampleValuesFromTheRFC(unittest.TestCase):
                 "algorithm": "SHA256",
             },
         )
-        self.assertEqual(hotp.provisioning_uri(), pyotp.parse_uri(hotp.provisioning_uri()).provisioning_uri())
+        self.assertEqual(hotp.provisioning_uri(), rlotp.parse_uri(hotp.provisioning_uri()).provisioning_uri())
 
-        hotp = pyotp.HOTP(key, digits=8, name="baco@peperina", issuer="Foo Corp", initial_count=10)
+        hotp = rlotp.HOTP(key, digits=8, name="baco@peperina", issuer="Foo Corp", initial_count=10)
         url = urlparse(hotp.provisioning_uri())
         self.assertEqual(url.scheme, "otpauth")
         self.assertEqual(url.netloc, "hotp")
@@ -95,15 +95,15 @@ class HOTPExampleValuesFromTheRFC(unittest.TestCase):
             dict(parse_qsl(url.query)),
             {"secret": "c7uxuqhgflpw7oruedmglbrk7u6242vb", "counter": "10", "issuer": "Foo Corp", "digits": "8"},
         )
-        self.assertEqual(hotp.provisioning_uri(), pyotp.parse_uri(hotp.provisioning_uri()).provisioning_uri())
+        self.assertEqual(hotp.provisioning_uri(), rlotp.parse_uri(hotp.provisioning_uri()).provisioning_uri())
 
-        code = pyotp.totp.TOTP("S46SQCPPTCNPROMHWYBDCTBZXV")
+        code = rlotp.totp.TOTP("S46SQCPPTCNPROMHWYBDCTBZXV")
         self.assertEqual(code.provisioning_uri(), "otpauth://totp/Secret?secret=S46SQCPPTCNPROMHWYBDCTBZXV")
         code.verify("123456")
         self.assertEqual(code.provisioning_uri(), "otpauth://totp/Secret?secret=S46SQCPPTCNPROMHWYBDCTBZXV")
 
     def test_other_secret(self):
-        hotp = pyotp.HOTP("N3OVNIBRERIO5OHGVCMDGS4V4RJ3AUZOUN34J6FRM4P6JIFCG3ZA")
+        hotp = rlotp.HOTP("N3OVNIBRERIO5OHGVCMDGS4V4RJ3AUZOUN34J6FRM4P6JIFCG3ZA")
         self.assertEqual(hotp.at(0), "737863")
         self.assertEqual(hotp.at(1), "390601")
         self.assertEqual(hotp.at(2), "363354")
@@ -141,7 +141,7 @@ class TOTPExampleValuesFromTheRFC(unittest.TestCase):
 
     def test_match_rfc(self):
         for digest, secret in self.RFC_VALUES:
-            totp = pyotp.TOTP(base64.b32encode(secret), 8, digest)
+            totp = rlotp.TOTP(base64.b32encode(secret), 8, digest)
             for utime, code in self.RFC_VALUES[(digest, secret)]:
                 if utime > sys.maxsize:
                     warn(
@@ -155,18 +155,18 @@ class TOTPExampleValuesFromTheRFC(unittest.TestCase):
                 self.assertEqual(value, str(code), msg)
 
     def test_match_rfc_digit_length(self):
-        totp = pyotp.TOTP("GEZDGNBVGY3TQOJQGEZDGNBVGY3TQOJQ")
+        totp = rlotp.TOTP("GEZDGNBVGY3TQOJQGEZDGNBVGY3TQOJQ")
         self.assertEqual(totp.at(1111111111), "050471")
         self.assertEqual(totp.at(1234567890), "005924")
         self.assertEqual(totp.at(2000000000), "279037")
 
     def test_match_google_authenticator_output(self):
-        totp = pyotp.TOTP("wrn3pqx5uqxqvnqr")
+        totp = rlotp.TOTP("wrn3pqx5uqxqvnqr")
         with Timecop(1297553958):
             self.assertEqual(totp.now(), "102705")
 
     def test_validate_totp(self):
-        totp = pyotp.TOTP("wrn3pqx5uqxqvnqr")
+        totp = rlotp.TOTP("wrn3pqx5uqxqvnqr")
         with Timecop(1297553958):
             self.assertTrue(totp.verify("102705"))
             self.assertTrue(totp.verify("102705"))
@@ -174,7 +174,7 @@ class TOTPExampleValuesFromTheRFC(unittest.TestCase):
             self.assertFalse(totp.verify("102705"))
 
     def test_input_before_epoch(self):
-        totp = pyotp.TOTP("GEZDGNBVGY3TQOJQGEZDGNBVGY3TQOJQ")
+        totp = rlotp.TOTP("GEZDGNBVGY3TQOJQGEZDGNBVGY3TQOJQ")
         # -1 and -29.5 round down to 0 (epoch)
         self.assertEqual(totp.at(-1), "755224")
         self.assertEqual(totp.at(-29.5), "755224")
@@ -182,31 +182,31 @@ class TOTPExampleValuesFromTheRFC(unittest.TestCase):
             totp.at(-30)
 
     def test_validate_totp_with_digit_length(self):
-        totp = pyotp.TOTP("GEZDGNBVGY3TQOJQGEZDGNBVGY3TQOJQ")
+        totp = rlotp.TOTP("GEZDGNBVGY3TQOJQGEZDGNBVGY3TQOJQ")
         with Timecop(1111111111):
             self.assertTrue(totp.verify("050471"))
         with Timecop(1297553958 + 30):
             self.assertFalse(totp.verify("050471"))
 
     def test_provisioning_uri(self):
-        totp = pyotp.TOTP("wrn3pqx5uqxqvnqr", name="mark@percival")
+        totp = rlotp.TOTP("wrn3pqx5uqxqvnqr", name="mark@percival")
         url = urlparse(totp.provisioning_uri())
         self.assertEqual(url.scheme, "otpauth")
         self.assertEqual(url.netloc, "totp")
         self.assertEqual(url.path, "/mark%40percival")
         self.assertEqual(dict(parse_qsl(url.query)), {"secret": "wrn3pqx5uqxqvnqr"})
-        self.assertEqual(totp.provisioning_uri(), pyotp.parse_uri(totp.provisioning_uri()).provisioning_uri())
+        self.assertEqual(totp.provisioning_uri(), rlotp.parse_uri(totp.provisioning_uri()).provisioning_uri())
 
-        totp = pyotp.TOTP("wrn3pqx5uqxqvnqr", name="mark@percival", issuer="FooCorp!")
+        totp = rlotp.TOTP("wrn3pqx5uqxqvnqr", name="mark@percival", issuer="FooCorp!")
         url = urlparse(totp.provisioning_uri())
         self.assertEqual(url.scheme, "otpauth")
         self.assertEqual(url.netloc, "totp")
         self.assertEqual(url.path, "/FooCorp%21:mark%40percival")
         self.assertEqual(dict(parse_qsl(url.query)), {"secret": "wrn3pqx5uqxqvnqr", "issuer": "FooCorp!"})
-        self.assertEqual(totp.provisioning_uri(), pyotp.parse_uri(totp.provisioning_uri()).provisioning_uri())
+        self.assertEqual(totp.provisioning_uri(), rlotp.parse_uri(totp.provisioning_uri()).provisioning_uri())
 
         key = "c7uxuqhgflpw7oruedmglbrk7u6242vb"
-        totp = pyotp.TOTP(key, digits=8, interval=60, digest=hashlib.sha256, name="baco@peperina", issuer="FooCorp")
+        totp = rlotp.TOTP(key, digits=8, interval=60, digest=hashlib.sha256, name="baco@peperina", issuer="FooCorp")
         url = urlparse(totp.provisioning_uri())
         self.assertEqual(url.scheme, "otpauth")
         self.assertEqual(url.netloc, "totp")
@@ -221,9 +221,9 @@ class TOTPExampleValuesFromTheRFC(unittest.TestCase):
                 "algorithm": "SHA256",
             },
         )
-        self.assertEqual(totp.provisioning_uri(), pyotp.parse_uri(totp.provisioning_uri()).provisioning_uri())
+        self.assertEqual(totp.provisioning_uri(), rlotp.parse_uri(totp.provisioning_uri()).provisioning_uri())
 
-        totp = pyotp.TOTP(key, digits=8, interval=60, name="baco@peperina", issuer="FooCorp")
+        totp = rlotp.TOTP(key, digits=8, interval=60, name="baco@peperina", issuer="FooCorp")
         url = urlparse(totp.provisioning_uri())
         self.assertEqual(url.scheme, "otpauth")
         self.assertEqual(url.netloc, "totp")
@@ -232,9 +232,9 @@ class TOTPExampleValuesFromTheRFC(unittest.TestCase):
             dict(parse_qsl(url.query)),
             {"secret": "c7uxuqhgflpw7oruedmglbrk7u6242vb", "issuer": "FooCorp", "digits": "8", "period": "60"},
         )
-        self.assertEqual(totp.provisioning_uri(), pyotp.parse_uri(totp.provisioning_uri()).provisioning_uri())
+        self.assertEqual(totp.provisioning_uri(), rlotp.parse_uri(totp.provisioning_uri()).provisioning_uri())
 
-        totp = pyotp.TOTP(key, digits=8, name="baco@peperina", issuer="FooCorp")
+        totp = rlotp.TOTP(key, digits=8, name="baco@peperina", issuer="FooCorp")
         url = urlparse(totp.provisioning_uri())
         self.assertEqual(url.scheme, "otpauth")
         self.assertEqual(url.netloc, "totp")
@@ -243,29 +243,29 @@ class TOTPExampleValuesFromTheRFC(unittest.TestCase):
             dict(parse_qsl(url.query)),
             {"secret": "c7uxuqhgflpw7oruedmglbrk7u6242vb", "issuer": "FooCorp", "digits": "8"},
         )
-        self.assertEqual(totp.provisioning_uri(), pyotp.parse_uri(totp.provisioning_uri()).provisioning_uri())
+        self.assertEqual(totp.provisioning_uri(), rlotp.parse_uri(totp.provisioning_uri()).provisioning_uri())
 
     def test_random_key_generation(self):
-        self.assertEqual(len(pyotp.random_base32()), 32)
-        self.assertEqual(len(pyotp.random_base32(length=34)), 34)
-        self.assertEqual(len(pyotp.random_hex()), 40)
-        self.assertEqual(len(pyotp.random_hex(length=42)), 42)
+        self.assertEqual(len(rlotp.random_base32()), 32)
+        self.assertEqual(len(rlotp.random_base32(length=34)), 34)
+        self.assertEqual(len(rlotp.random_hex()), 40)
+        self.assertEqual(len(rlotp.random_hex(length=42)), 42)
         with self.assertRaises(ValueError):
-            pyotp.random_base32(length=31)
+            rlotp.random_base32(length=31)
         with self.assertRaises(ValueError):
-            pyotp.random_hex(length=39)
+            rlotp.random_hex(length=39)
 
 
 class SteamTOTP(unittest.TestCase):
     def test_match_examples(self):
-        steam = pyotp.contrib.Steam("BASE32SECRET3232")
+        steam = rlotp.contrib.Steam("BASE32SECRET3232")
 
         self.assertEqual(steam.at(0), "2TC8B")
         self.assertEqual(steam.at(30), "YKKK4")
         self.assertEqual(steam.at(60), "M4HQB")
         self.assertEqual(steam.at(90), "DTVB3")
 
-        steam = pyotp.contrib.Steam("FMXNK4QEGKVPULRTADY6JIDK5VHUBGZW")
+        steam = rlotp.contrib.Steam("FMXNK4QEGKVPULRTADY6JIDK5VHUBGZW")
 
         self.assertEqual(steam.at(0), "C5V56")
         self.assertEqual(steam.at(30), "QJY8Y")
@@ -273,7 +273,7 @@ class SteamTOTP(unittest.TestCase):
         self.assertEqual(steam.at(90), "JG3T3")
 
     def test_verify(self):
-        steam = pyotp.contrib.Steam("BASE32SECRET3232")
+        steam = rlotp.contrib.Steam("BASE32SECRET3232")
         with Timecop(1662883100):
             self.assertTrue(steam.verify("N3G63"))
         with Timecop(1662883100 + 30):
@@ -284,7 +284,7 @@ class SteamTOTP(unittest.TestCase):
         with Timecop(946681223 + 30):
             self.assertFalse(steam.verify("7VP3X"))
 
-        steam = pyotp.contrib.Steam("FMXNK4QEGKVPULRTADY6JIDK5VHUBGZW")
+        steam = rlotp.contrib.Steam("FMXNK4QEGKVPULRTADY6JIDK5VHUBGZW")
         with Timecop(1662884261):
             self.assertTrue(steam.verify("V6WKJ"))
         with Timecop(1662884261 + 30):
@@ -297,7 +297,7 @@ class SteamTOTP(unittest.TestCase):
 
 
 class CompareDigestTest(unittest.TestCase):
-    method = staticmethod(pyotp.utils.compare_digest)
+    method = staticmethod(rlotp.utils.compare_digest)
 
     def test_comparisons(self):
         self.assertTrue(self.method("", ""))
@@ -310,7 +310,7 @@ class CompareDigestTest(unittest.TestCase):
 
 
 class StringComparisonTest(CompareDigestTest):
-    method = staticmethod(pyotp.utils.strings_equal)
+    method = staticmethod(rlotp.utils.strings_equal)
 
     def test_fullwidth_input(self):
         self.assertTrue(self.method("ｘs１２３45", "xs12345"))
@@ -321,14 +321,14 @@ class StringComparisonTest(CompareDigestTest):
 
 class CounterOffsetTest(unittest.TestCase):
     def test_counter_offset(self):
-        totp = pyotp.TOTP("ABCDEFGH")
+        totp = rlotp.TOTP("ABCDEFGH")
         self.assertEqual(totp.at(200), "028307")
         self.assertTrue(totp.at(200, 1), "681610")
 
 
 class ValidWindowTest(unittest.TestCase):
     def test_valid_window(self):
-        totp = pyotp.TOTP("ABCDEFGH")
+        totp = rlotp.TOTP("ABCDEFGH")
         self.assertTrue(totp.verify("451564", 200, 1))
         self.assertTrue(totp.verify("028307", 200, 1))
         self.assertTrue(totp.verify("681610", 200, 1))
@@ -338,43 +338,43 @@ class ValidWindowTest(unittest.TestCase):
 class ParseUriTest(unittest.TestCase):
     def test_invalids(self):
         with self.assertRaises(ValueError) as cm:
-            pyotp.parse_uri("http://hello.com")
+            rlotp.parse_uri("http://hello.com")
         self.assertEqual("Not an otpauth URI", str(cm.exception))
 
         with self.assertRaises(ValueError) as cm:
-            pyotp.parse_uri("otpauth://totp")
+            rlotp.parse_uri("otpauth://totp")
         self.assertEqual("No secret found in URI", str(cm.exception))
 
         with self.assertRaises(ValueError) as cm:
-            pyotp.parse_uri("otpauth://derp?secret=foo")
+            rlotp.parse_uri("otpauth://derp?secret=foo")
         self.assertEqual("Not a supported OTP type", str(cm.exception))
 
         with self.assertRaises(ValueError) as cm:
-            pyotp.parse_uri("otpauth://totp?foo=secret")
+            rlotp.parse_uri("otpauth://totp?foo=secret")
         self.assertEqual("foo is not a valid parameter", str(cm.exception))
 
         with self.assertRaises(ValueError) as cm:
-            pyotp.parse_uri("otpauth://totp?digits=-1")
+            rlotp.parse_uri("otpauth://totp?digits=-1")
         self.assertEqual("Digits may only be 6, 7, or 8", str(cm.exception))
 
         with self.assertRaises(ValueError) as cm:
-            pyotp.parse_uri("otpauth://totp/SomeIssuer:?issuer=AnotherIssuer")
+            rlotp.parse_uri("otpauth://totp/SomeIssuer:?issuer=AnotherIssuer")
         self.assertEqual("If issuer is specified in both label and parameters, it should be equal.", str(cm.exception))
 
         with self.assertRaises(ValueError) as cm:
-            pyotp.parse_uri("otpauth://totp?algorithm=aes")
+            rlotp.parse_uri("otpauth://totp?algorithm=aes")
         self.assertEqual("Invalid value for algorithm, must be SHA1, SHA256 or SHA512", str(cm.exception))
-    
-    def test_parse_steam(self):
-        otp = pyotp.parse_uri("otpauth://totp/Steam:?secret=SOME_SECRET&encoder=steam")
-        self.assertEqual(type(otp), pyotp.contrib.Steam)
 
-        otp = pyotp.parse_uri("otpauth://totp/Steam:?secret=SOME_SECRET")
-        self.assertNotEqual(type(otp), pyotp.contrib.Steam)
+    def test_parse_steam(self):
+        otp = rlotp.parse_uri("otpauth://totp/Steam:?secret=SOME_SECRET&encoder=steam")
+        self.assertEqual(type(otp), rlotp.contrib.Steam)
+
+        otp = rlotp.parse_uri("otpauth://totp/Steam:?secret=SOME_SECRET")
+        self.assertNotEqual(type(otp), rlotp.contrib.Steam)
 
     @unittest.skipIf(sys.version_info < (3, 6), "Skipping test that requires deterministic dict key enumeration")
     def test_algorithms(self):
-        otp = pyotp.parse_uri("otpauth://totp?algorithm=SHA1&secret=GEZDGNBV&algorithm=SHA1")
+        otp = rlotp.parse_uri("otpauth://totp?algorithm=SHA1&secret=GEZDGNBV&algorithm=SHA1")
         self.assertEqual(hashlib.sha1, otp.digest)
         self.assertEqual(otp.at(0), "734055")
         self.assertEqual(otp.at(30), "662488")
@@ -382,7 +382,7 @@ class ParseUriTest(unittest.TestCase):
         self.assertEqual(otp.provisioning_uri(), "otpauth://totp/Secret?secret=GEZDGNBV")
         self.assertEqual(otp.provisioning_uri(name="n", issuer_name="i"), "otpauth://totp/i:n?secret=GEZDGNBV&issuer=i")
 
-        otp = pyotp.parse_uri("otpauth://totp?algorithm=SHA1&secret=GEZDGNBV&algorithm=SHA1&period=60")
+        otp = rlotp.parse_uri("otpauth://totp?algorithm=SHA1&secret=GEZDGNBV&algorithm=SHA1&period=60")
         self.assertEqual(hashlib.sha1, otp.digest)
         self.assertEqual(otp.at(30), "734055")
         self.assertEqual(otp.at(60), "662488")
@@ -390,7 +390,7 @@ class ParseUriTest(unittest.TestCase):
             otp.provisioning_uri(name="n", issuer_name="i"), "otpauth://totp/i:n?secret=GEZDGNBV&issuer=i&period=60"
         )
 
-        otp = pyotp.parse_uri("otpauth://hotp?algorithm=SHA1&secret=GEZDGNBV&algorithm=SHA1")
+        otp = rlotp.parse_uri("otpauth://hotp?algorithm=SHA1&secret=GEZDGNBV&algorithm=SHA1")
         self.assertEqual(hashlib.sha1, otp.digest)
         self.assertEqual(otp.at(0), "734055")
         self.assertEqual(otp.at(1), "662488")
@@ -399,7 +399,7 @@ class ParseUriTest(unittest.TestCase):
             otp.provisioning_uri(name="n", issuer_name="i"), "otpauth://hotp/i:n?secret=GEZDGNBV&issuer=i&counter=0"
         )
 
-        otp = pyotp.parse_uri("otpauth://hotp?algorithm=SHA1&secret=GEZDGNBV&algorithm=SHA1&counter=1")
+        otp = rlotp.parse_uri("otpauth://hotp?algorithm=SHA1&secret=GEZDGNBV&algorithm=SHA1&counter=1")
         self.assertEqual(hashlib.sha1, otp.digest)
         self.assertEqual(otp.at(0), "662488")
         self.assertEqual(otp.at(1), "289363")
@@ -407,12 +407,12 @@ class ParseUriTest(unittest.TestCase):
             otp.provisioning_uri(name="n", issuer_name="i"), "otpauth://hotp/i:n?secret=GEZDGNBV&issuer=i&counter=1"
         )
 
-        otp = pyotp.parse_uri("otpauth://totp?algorithm=SHA1&secret=GEZDGNBV&algorithm=SHA256")
+        otp = rlotp.parse_uri("otpauth://totp?algorithm=SHA1&secret=GEZDGNBV&algorithm=SHA256")
         self.assertEqual(hashlib.sha256, otp.digest)
         self.assertEqual(otp.at(0), "918961")
         self.assertEqual(otp.at(9000), "934470")
 
-        otp = pyotp.parse_uri("otpauth://totp?algorithm=SHA1&secret=GEZDGNBV&algorithm=SHA512")
+        otp = rlotp.parse_uri("otpauth://totp?algorithm=SHA1&secret=GEZDGNBV&algorithm=SHA512")
         self.assertEqual(hashlib.sha512, otp.digest)
         self.assertEqual(otp.at(0), "816660")
         self.assertEqual(otp.at(9000), "524153")
@@ -424,19 +424,19 @@ class ParseUriTest(unittest.TestCase):
         with self.assertRaises(ValueError):
             otp.provisioning_uri(name="n", issuer_name="i", image="nourl")
 
-        otp = pyotp.parse_uri(otp.provisioning_uri(name="n", issuer_name="i", image="https://test.net/test.png"))
+        otp = rlotp.parse_uri(otp.provisioning_uri(name="n", issuer_name="i", image="https://test.net/test.png"))
         self.assertEqual(hashlib.sha512, otp.digest)
 
-        otp = pyotp.parse_uri("otpauth://totp/Steam:?secret=FMXNK4QEGKVPULRTADY6JIDK5VHUBGZW&encoder=steam")
-        self.assertEqual(type(otp), pyotp.contrib.Steam)
+        otp = rlotp.parse_uri("otpauth://totp/Steam:?secret=FMXNK4QEGKVPULRTADY6JIDK5VHUBGZW&encoder=steam")
+        self.assertEqual(type(otp), rlotp.contrib.Steam)
         self.assertEqual(otp.at(0), "C5V56")
         self.assertEqual(otp.at(30), "QJY8Y")
         self.assertEqual(otp.at(60), "R3WQY")
         self.assertEqual(otp.at(90), "JG3T3")
 
         # period and digits should be ignored
-        otp = pyotp.parse_uri("otpauth://totp/Steam:?secret=FMXNK4QEGKVPULRTADY6JIDK5VHUBGZW&period=15&digits=7&encoder=steam")
-        self.assertEqual(type(otp), pyotp.contrib.Steam)
+        otp = rlotp.parse_uri("otpauth://totp/Steam:?secret=FMXNK4QEGKVPULRTADY6JIDK5VHUBGZW&period=15&digits=7&encoder=steam")
+        self.assertEqual(type(otp), rlotp.contrib.Steam)
         self.assertEqual(otp.at(0), "C5V56")
         self.assertEqual(otp.at(30), "QJY8Y")
         self.assertEqual(otp.at(60), "R3WQY")